+++ Tests/MapboxMapsTests/Foundation/Camera/MockPropertyAnimator.swift
@@ -5,54 +5,41 @@ final class MockPropertyAnimator: UIViewPropertyAnimator {
     let setIsReversedStub = Stub<Bool, Void>()
     override var isReversed: Bool {
         get {
-            fatalError("unimplemented")
+            return super.isReversed
         }
         set {
+            super.isReversed = newValue
             setIsReversedStub.call(with: newValue)
         }
     }
 
-    let stateStub = Stub<Void, UIViewAnimatingState>(defaultReturnValue: .inactive)
-    override var state: UIViewAnimatingState {
-        stateStub.call()
-    }
-
     let startAnimationStub = Stub<Void, Void>()
     override func startAnimation() {
+        super.startAnimation()
         startAnimationStub.call()
     }
 
     let stopAnimationStub = Stub<Bool, Void>()
     override func stopAnimation(_ withoutFinishing: Bool) {
+        super.stopAnimation(withoutFinishing)
         stopAnimationStub.call(with: withoutFinishing)
-    }
-
-    let pauseAnimationsStub = Stub<Void, Void>()
-    override func pauseAnimation() {
-        pauseAnimationsStub.call()
     }
 
     let addAnimationsStub = Stub<() -> Void, Void>()
     override func addAnimations(_ animation: @escaping () -> Void) {
+        super.addAnimations(animation)
         addAnimationsStub.call(with: animation)
     }
 
     let addCompletionStub = Stub<(UIViewAnimatingPosition) -> Void, Void>()
     override func addCompletion(_ completion: @escaping (UIViewAnimatingPosition) -> Void) {
+        super.addCompletion(completion)
         addCompletionStub.call(with: completion)
-    }
-
-    struct ContinueAnimationParameters {
-        var parameters: UITimingCurveProvider?
-        var durationFactor: CGFloat
-    }
-    let continueAnimationStub = Stub<ContinueAnimationParameters, Void>()
-    override func continueAnimation(withTimingParameters parameters: UITimingCurveProvider?, durationFactor: CGFloat) {
-        continueAnimationStub.call(with: .init(parameters: parameters, durationFactor: durationFactor))
     }
 
     let finishAnimationStub = Stub<UIViewAnimatingPosition, Void>()
     override func finishAnimation(at finalPosition: UIViewAnimatingPosition) {
+        super.finishAnimation(at: finalPosition)
         finishAnimationStub.call(with: finalPosition)
     }
 }