+++ CHANGELOG.md
@@ -4,8 +4,9 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## main
 
-* Make PointAnnotation.Image's fields public. ([#753](https://github.com/mapbox/mapbox-maps-ios/pull/753))
-* Fixes a bug where unsupported locale's were crashing. ([#752](https://github.com/mapbox/mapbox-maps-ios/pull/752))
+* Make `PointAnnotation.Image`'s fields public. ([#753](https://github.com/mapbox/mapbox-maps-ios/pull/753))
+* Passing an unsupported locale into `Style.localizeLabels(into:forLayerIds:)` throws an error instead of crashing. ([#752](https://github.com/mapbox/mapbox-maps-ios/pull/752))
+* Set `MapboxMap` flags during gestures and animations. ([#754](https://github.com/mapbox/mapbox-maps-ios/pull/754))
 
 ## 10.0.0 - October 6, 2021